Update rounding decimal converter to check for decimal not double

+++ src/Vendr.Deploy/Converters/RoundingDecimalJsonConverter.cs
@@ -26,7 +26,7 @@ namespace Vendr.Deploy.Converters
             => false;
 
         public override bool CanConvert(Type objectType)
-            => objectType == typeof(double);
+            => objectType == typeof(decimal);
 
         public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
             => throw new NotImplementedException();
